remove extra file
