+++ packages/fuels-programs/src/call_utils.rs
@@ -10,7 +10,7 @@ use fuels_core::{
     offsets::call_script_data_offset,
     types::{
         bech32::Bech32Address,
-        errors::Result,
+        errors::{Error as FuelsError, Result},
         input::Input,
         param_types::ParamType,
         transaction::{ScriptTransaction, TxParameters},
@@ -29,6 +29,36 @@ pub(crate) struct CallOpcodeParamsOffset {
     pub amount_offset: usize,
     pub gas_forwarded_offset: usize,
     pub call_data_offset: usize,
+}
+
+/// How many times to attempt to resolve missing tx dependencies.
+pub const DEFAULT_TX_DEP_ESTIMATION_ATTEMPTS: u64 = 10;
+
+#[async_trait::async_trait]
+pub trait TxDependencyEstimation {
+    async fn simulate(&mut self) -> Result<()>;
+    fn append_missing_deps(self, receipts: &[Receipt]) -> Self;
+}
+
+pub async fn estimate_tx_dependencies<T: TxDependencyEstimation + Send + Sync>(
+    mut program_handler: T,
+    max_attempts: Option<u64>,
+) -> Result<T> {
+    let attempts = max_attempts.unwrap_or(DEFAULT_TX_DEP_ESTIMATION_ATTEMPTS);
+
+    for _ in 0..attempts {
+        match program_handler.simulate().await {
+            Ok(_) => return Ok(program_handler),
+
+            Err(FuelsError::RevertTransactionError { ref receipts, .. }) => {
+                program_handler = program_handler.append_missing_deps(receipts);
+            }
+
+            Err(other_error) => return Err(other_error),
+        }
+    }
+
+    program_handler.simulate().await.map(|_| program_handler)
 }
 
 /// Creates a [`ScriptTransaction`] from contract calls. The internal [Transaction] is
