+++ packages/fuels-programs/src/contract.rs
@@ -25,11 +25,11 @@ use itertools::Itertools;
 use crate::{
     call_response::FuelCallResponse,
     call_utils::{
-        build_tx_from_contract_calls, find_contract_not_in_inputs, is_missing_output_variables,
+        build_tx_from_contract_calls, estimate_tx_dependencies, find_contract_not_in_inputs,
+        is_missing_output_variables, TxDependencyEstimation,
     },
     logs::{map_revert_error, LogDecoder},
     receipt_parser::ReceiptParser,
-    tx_dependency_estimation::{TxDependencyEstimation, TxDependencyEstimator},
 };
 
 #[derive(Debug, Clone)]
@@ -555,7 +555,7 @@ where
     /// Simulates the call and attempts to resolve missing tx dependencies.
     /// Forwards the received error if it cannot be fixed.
     pub async fn estimate_tx_dependencies(self, max_attempts: Option<u64>) -> Result<Self> {
-        TxDependencyEstimator::estimate_tx_dependencies(self, max_attempts).await
+        estimate_tx_dependencies(self, max_attempts).await
     }
 
     /// Get a contract's estimated cost
@@ -797,7 +797,7 @@ impl<T: Account> MultiContractCallHandler<T> {
     /// Simulates the call and attempts to resolve missing tx dependencies.
     /// Forwards the received error if it cannot be fixed.
     pub async fn estimate_tx_dependencies(self, max_attempts: Option<u64>) -> Result<Self> {
-        TxDependencyEstimator::estimate_tx_dependencies(self, max_attempts).await
+        estimate_tx_dependencies(self, max_attempts).await
     }
 
     /// Get a contract's estimated cost
