+++ packages/fuels-programs/src/script_calls.rs
@@ -24,13 +24,12 @@ use itertools::chain;
 use crate::{
     call_response::FuelCallResponse,
     call_utils::{
-        find_contract_not_in_inputs, generate_contract_inputs, generate_contract_outputs,
-        is_missing_output_variables,
+        estimate_tx_dependencies, find_contract_not_in_inputs, generate_contract_inputs,
+        generate_contract_outputs, is_missing_output_variables, TxDependencyEstimation,
     },
     contract::SettableContract,
     logs::{map_revert_error, LogDecoder},
     receipt_parser::ReceiptParser,
-    tx_dependency_estimation::{TxDependencyEstimation, TxDependencyEstimator},
 };
 
 #[derive(Debug)]
@@ -306,7 +305,7 @@ where
     /// Simulates the call and attempts to resolve missing tx dependencies.
     /// Forwards the received error if it cannot be fixed.
     pub async fn estimate_tx_dependencies(self, max_attempts: Option<u64>) -> Result<Self> {
-        TxDependencyEstimator::estimate_tx_dependencies(self, max_attempts).await
+        estimate_tx_dependencies(self, max_attempts).await
     }
 
     /// Create a [`FuelCallResponse`] from call receipts
