+++ packages/fuels/src/lib.rs
@@ -84,7 +84,6 @@ pub mod prelude {
                 SettableContract, StorageConfiguration,
             },
             logs::{LogDecoder, LogId, LogResult},
-            tx_dependency_estimation::TxDependencyEstimation,
         },
         test_helpers::*,
     };
